requirements.txt
Flask
Flask-CORS
PyJWT
Werkzeug
